refactor(monitoring): extract execution health signals from health_checks
health_checks.py (1279 lines) mixed broker/ticker probing and the
HealthCheckRunner with a self-contained cluster of pure functions that derive
execution health signals (order error/retry rate, broker latency p95, guard
trips, missing decision ids) from the metrics collector summary. Move that
cluster into monitoring/execution_health_signals.py (1279->1097) and re-export
compute_execution_health_signals.

The cluster is pure (counter/histogram math, local health_signals imports) and
referenced nothing in the parent; the entangled staleness checks
(_resolve_last_ticker_timestamp dependency) intentionally stay. Re-export keeps
the public name stable for health_server/status_reporter; 602 monitoring tests pass.
